implement suggestions by Julien
